move certificates frontend code into a new subdirectory
