+++ apps/src/sites/studio/pages/congrats/index.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Congrats from '@cdo/apps/templates/Congrats';
+import Congrats from '@cdo/apps/templates/certificates/Congrats';
 import {Provider} from 'react-redux';
 import {getStore} from '@cdo/apps/redux';
 import queryString from 'query-string';
