+++ apps/test/unit/templates/StudentsBeyondHocTest.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {Provider} from 'react-redux';
 import {mount} from 'enzyme';
 import {expect} from '../../util/reconfiguredChai';
-import StudentsBeyondHoc from '@cdo/apps/templates/StudentsBeyondHoc';
+import StudentsBeyondHoc from '@cdo/apps/templates/certificates/StudentsBeyondHoc';
 import {combineReducers, createStore} from 'redux';
 import responsive from '@cdo/apps/code-studio/responsiveRedux';
 import isRtl from '@cdo/apps/code-studio/isRtlRedux';
